requirements.txt
anyio==3.6.1
cryptography==37.0.4
fastapi==0.79.0
motor==3.0.0
pydantic==1.9.2
pymongo[srv]==4.2.0
python-dotenv==0.20.0
python-jose==3.3.0
starlette==0.19.1
uvicorn==0.18.2
loguru==0.6.0
aiohttp~=3.8.1
aioredis~=2.0.1